@@ -1,98 +1,108 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using System.Security.Claims;
+using Microsoft.Extensions.Hosting;
 using SIFCore.Models;
-using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+using SIFCore.Services;
 
 namespace SIFCore
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
+        public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
             Configuration = configuration;
+            Env = env;
         }
 
         public IConfiguration Configuration { get; }
+        public IWebHostEnvironment Env { get; set; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-             services.AddMvc();
-
-             services.AddDbContext<SIFContext>();
+            services.AddScoped<IIdentityService, IdentityService>();
             
-            services.AddAuthentication( "Cookies") // Sets the default scheme to cookies
-                .AddCookie( "Cookies", options =>
+            services.AddControllersWithViews();  
+
+            IMvcBuilder builder = services.AddRazorPages(); 
+
+            #if DEBUG
+                if (Env.IsDevelopment())
+                {
+                    builder.AddRazorRuntimeCompilation();
+                }
+            #endif        
+
+            services.AddDbContext<SIFContext>();
+      
+            services.AddAuthentication("Cookies") // Sets the default scheme to cookies
+                .AddCookie("Cookies", options =>
                 {
                     options.AccessDeniedPath = "/account/denied";
                     options.LoginPath = "/account/login";
                     options.ExpireTimeSpan = TimeSpan.FromMinutes(30);
-                    options.SlidingExpiration = true;                    
-                });                
-                // .AddCAS(o =>
-                // {
-                //     o.CasServerUrlBase = Configuration["CasBaseUrl"];   // Set in `appsettings.json` file.
-                //     o.SignInScheme = "Cookies";
-                //     o.Events.OnTicketReceived = async context => {
-                //         var identity = (ClaimsIdentity) context.Principal.Identity;
-                //         if (identity == null)
-                //         {
-                //             return;
-                //         }
+                    options.SlidingExpiration = true;
 
-                //         // kerb comes across in name & name identifier
-                //         var kerb = identity?.FindFirst(ClaimTypes.NameIdentifier).Value;
+                })
+                .AddCAS(o =>
+                {
+                    o.CasServerUrlBase = Configuration["CasBaseUrl"];   // Set in `appsettings.json` file.
+                    o.SignInScheme = "Cookies";
+                    o.Events.OnCreatingTicket = async context => {
+                        var identity = (ClaimsIdentity) context.Principal.Identity;
+                        var assertion = context.Assertion;
+                        if (identity == null)
+                        {
+                            return;
+                        }
 
-                //         if (string.IsNullOrWhiteSpace(kerb)) return;
+                        var kerb = assertion.PrincipalName;
 
-                //         var identityService = services.BuildServiceProvider().GetService<IIdentityService>();
+                        if (string.IsNullOrWhiteSpace(kerb)) return;
 
-                //         var user = await identityService.GetByKerberos(kerb);
+                        var identityService = services.BuildServiceProvider().GetService<IIdentityService>();
+                        
 
-                //         if (user == null)
-                //         {
-                //             return;
-                //         }                        
+                        var user = await identityService.GetByKerberos(kerb);
 
-                //         identity.RemoveClaim(identity.FindFirst(ClaimTypes.Name));
-                //         identity.AddClaim(new Claim(ClaimTypes.Name, user.Id));
+                        if (user == null)
+                        {
+                            return;
+                        }                        
 
-                //         identity.AddClaim(new Claim(ClaimTypes.GivenName, user.FirstName));
-                //         identity.AddClaim(new Claim(ClaimTypes.Surname, user.LastName));
-                //         identity.AddClaim(new Claim("name", user.Name));
-                //         identity.AddClaim(new Claim(ClaimTypes.Email, user.Email));
-                //         identity.RemoveClaim(identity.FindFirst(ClaimTypes.NameIdentifier));
-                //         identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, kerb));
-                //         identity.AddClaim(new Claim(ClaimTypes.Role, "Employee"));
-                //         if(!user.SeasonalEmployee)
-                //         {
-                //             identity.AddClaim(new Claim(ClaimTypes.Role, "AllowEmulate"));
-                //         }
+                         var existingClaim = identity.FindFirst(ClaimTypes.Name);
+                        if(existingClaim != null)
+                        {
+                            identity.RemoveClaim(identity.FindFirst(ClaimTypes.Name));
+                        }   
+                        identity.AddClaim(new Claim(ClaimTypes.Name, user.Id.ToString()));
 
-                //         context.Principal.AddIdentity(identity);
+                        identity.AddClaim(new Claim(ClaimTypes.GivenName, user.FirstName));
+                        identity.AddClaim(new Claim(ClaimTypes.Surname, user.LastName));
+                        identity.AddClaim(new Claim("name", user.FullName));
+                        identity.AddClaim(new Claim(ClaimTypes.Email, user.Email));
+                        existingClaim = identity.FindFirst(ClaimTypes.NameIdentifier);
+                        if(existingClaim != null)
+                        {
+                            identity.RemoveClaim(identity.FindFirst(ClaimTypes.NameIdentifier));
+                        }
+                        identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, kerb));
+                        identity.AddClaim(new Claim(ClaimTypes.Role, "Employee"));                        
 
-                //         await Task.FromResult(0); 
-                //     };
-                // });
+                        context.Principal.AddIdentity(identity);
 
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                        await Task.FromResult(0); 
+                    };
+                });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, Microsoft.Extensions.Hosting.IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -103,30 +113,33 @@ namespace SIFCore
                 app.UseExceptionHandler("/Home/Error");
                 // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
-            }        
-
+            }
             app.UseHttpsRedirection();
             app.UseStaticFiles();
             app.UseCookiePolicy();
-            app.UseAuthentication(); 
+            app.UseAuthentication();
 
-            app.UseMvc(routes =>
+            app.UseRouting();
+
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapAreaRoute(
-                   name: "Client_route",
+                endpoints.MapAreaControllerRoute(
+                    name: "Client_route",
                    areaName: "Client",
-                   template:  "client/{controller}/{action=Index}/{id?}"
-               );
+                   pattern:  "client/{controller}/{action=Index}/{id?}"
+                );
 
-               routes.MapAreaRoute(
-                   name: "Admin_route",
+                endpoints.MapAreaControllerRoute(
+                    name: "Admin_route",
                    areaName: "Admin",
-                   template:  "admin/{controller}/{action=Index}/{id?}"
-               );
-               
-                routes.MapRoute(
+                   pattern:  "admin/{controller}/{action=Index}/{id?}"
+                );
+
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
