@@ -10,9 +10,9 @@ namespace SIFCore.Models
      public class CreateOrdersViewModel
 	{
 		public Orders Order { get; set; }
-       // public List<Analysis> Analysises { get; set; }
-        // public List<AnalysisTypes> AnalysisTypes { get; set; }
-        // public List<Requirements> Requirements { get; set; }
+        public List<Analysis> Analyses { get; set; }
+        public List<AnalysisTypes> AnalysisTypes { get; set; }
+        public List<Requirements> Requirements { get; set; }
         public List<ShippingAddresses> ShippingAddresses { get; set; }
 	    public List<BillingAddresses>   BillingAddresses { get; set; }
         public BillingAddresses Billing { get; set; }
@@ -25,9 +25,6 @@ namespace SIFCore.Models
             var viewModel = new CreateOrdersViewModel
             {
                 Order = await _dbContext.Orders.Where(o => o.Id == orderId).FirstOrDefaultAsync(),
-                // Analysises = await _dbContext.Analysis.Where(a => a.OrderId == orderId).ToListAsync(),
-                // AnalysisTypes = await _dbContext.AnalysisTypes.OrderBy(at => at.AnalysisOrder).ToListAsync(),
-                // Requirements = await _dbContext.Requirements.Where(r => r.CurrentAnalysis).OrderBy(r => r.AnalysisOrder).ToListAsync(),
                 ShippingAddresses = await _dbContext.ShippingAddresses.Where(s => s.ContactId.ToString() == contactId).OrderBy(s => s.AddressName).ToListAsync(),
                 BillingAddresses = await _dbContext.BillingAddresses.Where(b => b.ContactId.ToString() == contactId).OrderBy(b => b.AddressName).ToListAsync(),
                 PaymentOptions = EnumHelper.GetListOfDisplayNames<PaymentTypes>(),
@@ -35,5 +32,24 @@ namespace SIFCore.Models
  
 			return viewModel;
 		}
+
+        public static async Task<CreateOrdersViewModel> EditViewModel(SIFContext _dbContext, int orderId, string contactId)
+        {
+            var viewModel = new CreateOrdersViewModel
+            {
+                Order = await _dbContext.Orders.Where(o => o.Id == orderId).FirstOrDefaultAsync(),
+                Analyses = await _dbContext.Analysis.Where(a => a.OrderId == orderId)
+                    .Include(a => a.AnalysisRequirement)
+                    .ToListAsync(),
+                AnalysisTypes = await _dbContext.AnalysisTypes.OrderBy(at => at.AnalysisOrder).ToListAsync(),
+                Requirements = await _dbContext.Requirements.Where(r => r.CurrentAnalysis).OrderBy(r => r.AnalysisOrder).ToListAsync(),
+                ShippingAddresses = await _dbContext.ShippingAddresses.Where(s => s.ContactId.ToString() == contactId).OrderBy(s => s.AddressName).ToListAsync(),
+                BillingAddresses = await _dbContext.BillingAddresses.Where(b => b.ContactId.ToString() == contactId).OrderBy(b => b.AddressName).ToListAsync(),
+                PaymentOptions = EnumHelper.GetListOfDisplayNames<PaymentTypes>(),
+            };
+ 
+			return viewModel;
+
+        }
 	}
 }
