@@ -14,67 +14,83 @@
     </div>
     <div class="form-group row required">
         <label asp-for="AddressName" class="col-sm-3 col-form-label"></label>        
-        <input asp-for="AddressName" class="col-sm-9 form-control">        
+        <input asp-for="AddressName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="AddressName" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="Institution" class="col-sm-3 col-form-label"></label>
         <input asp-for="Institution" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Institution" data-valmsg-replace="true"></span>
     </div>    
     <div class="form-group row">
         <label asp-for="Department" class="col-sm-3 col-form-label"></label>
         <input asp-for="Department" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Department" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="ContactFirstName" class="col-sm-3 col-form-label"></label>
         <input asp-for="ContactFirstName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ContactFirstName" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="ContactLastName" class="col-sm-3 col-form-label"></label>
        <input asp-for="ContactLastName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ContactLastName" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="BillingEmailAddress" class="col-sm-3 col-form-label"></label>
         <input asp-for="BillingEmailAddress" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="BillingEmailAddress" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="BillingFax" class="col-sm-3 col-form-label"></label>
         <input asp-for="BillingFax" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="BillingFax" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="BillingPhone" class="col-sm-3 col-form-label"></label>
         <input asp-for="BillingPhone" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="BillingPhone" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="Address1" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address1" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address1" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="Address2" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address2" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address2" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="Address3" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address3" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address3" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="City" class="col-sm-3 col-form-label"></label>
         <input asp-for="City" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="City" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="State" class="col-sm-3 col-form-label"></label>
         <input asp-for="State" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="State" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="ZipCode" class="col-sm-3 col-form-label"></label>
         <input asp-for="ZipCode" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ZipCode" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="Country" class="col-sm-3 col-form-label"></label>
         <input asp-for="Country" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Country" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="FedIDNum" class="col-sm-3 col-form-label"></label>
         <input asp-for="FedIDNum" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="FedIDNum" data-valmsg-replace="true"></span>
     </div>
     <div class="row">
         <input type="submit" value="Update" class="btn btn-primary"> | 
