@@ -1,7 +1,7 @@
 @model SIFCore.Models.ShippingAddresses;
 
 
-<h2>Shipping Details</h2> 
+<h2>Data Report Contact Details</h2> 
 <div class="container col-lg-5"> 
     <div class="row">
         <a asp-action="ShippingEdit" class="btn btn-primary" asp-route-id="@Model.Id">Edit</a> | 
@@ -42,13 +42,8 @@
         <div class="col-sm-9 form-control-plaintext">
             @Html.DisplayFor(x => x.PIEmail)
         </div>
-    </div>
+    </div>    
     <div class="form-group row">
-        <label asp-for="PIEmailAdditional" class="col-sm-3 col-form-label"></label>
-        <div class="col-sm-9 form-control-plaintext">
-            @Html.DisplayFor(x => x.PIEmailAdditional)
-        </div>
-    </div><div class="form-group row">
         <label asp-for="PIFax" class="col-sm-3 col-form-label"></label>
         <div class="col-sm-9 form-control-plaintext">
             @Html.DisplayFor(x => x.PIFax)
@@ -58,6 +53,30 @@
         <label asp-for="PIPhone" class="col-sm-3 col-form-label"></label>
         <div class="col-sm-9 form-control-plaintext">
             @Html.DisplayFor(x => x.PIPhone)
+        </div>
+    </div>
+    <div class="form-group row">
+        <label asp-for="ResearcherFirstName" class="col-sm-3 col-form-label"></label>
+        <div class="col-sm-9 form-control-plaintext">
+            @Html.DisplayFor(x => x.ResearcherFirstName)
+        </div>
+    </div>
+    <div class="form-group row">
+        <label asp-for="ResearcherLastName" class="col-sm-3 col-form-label"></label>
+        <div class="col-sm-9 form-control-plaintext">
+            @Html.DisplayFor(x => x.ResearcherLastName)
+        </div>
+    </div>
+    <div class="form-group row">
+        <label asp-for="ResearcherEmail" class="col-sm-3 col-form-label"></label>
+        <div class="col-sm-9 form-control-plaintext">
+            @Html.DisplayFor(x => x.ResearcherEmail)
+        </div>
+    </div>
+    <div class="form-group row">
+        <label asp-for="ResearcherPhone" class="col-sm-3 col-form-label"></label>
+        <div class="col-sm-9 form-control-plaintext">
+            @Html.DisplayFor(x => x.ResearcherPhone)
         </div>
     </div>
     <div class="form-group row">
@@ -100,6 +119,12 @@
         <label asp-for="Country" class="col-sm-3 col-form-label"></label>
         <div class="col-sm-9 form-control-plaintext">
             @Html.DisplayFor(x => x.Country)
+        </div>
+    </div>
+    <div class="form-group row">
+        <label asp-for="Notes" class="col-sm-3 col-form-label"></label>
+        <div class="col-sm-9 form-control-plaintext">
+           @Html.DisplayFor(x => x.Notes)
         </div>
     </div>
     <div class="row">
