@@ -1,7 +1,7 @@
 @model SIFCore.Models.ShippingAddresses;
 
 
-<h2>Shipping Address Edit</h2> 
+<h2>Data Report Contact Edit</h2> 
 <form asp-action="ShippingEdit">
     @Html.HiddenFor(x => x.Id)
 <div class="container col-lg-5">     
@@ -14,72 +14,110 @@
     </div>
     <div class="form-group row required">
         <label asp-for="AddressName" class="col-sm-3 col-form-label"></label>        
-        <input asp-for="AddressName" class="col-sm-9 form-control">        
+        <input asp-for="AddressName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="AddressName" data-valmsg-replace="true"></span>     
     </div>
     <div class="form-group row">
         <label asp-for="Institution" class="col-sm-3 col-form-label"></label>
         <input asp-for="Institution" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Institution" data-valmsg-replace="true"></span>
     </div>    
     <div class="form-group row">
         <label asp-for="Department" class="col-sm-3 col-form-label"></label>
         <input asp-for="Department" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Department" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="PIFirstName" class="col-sm-3 col-form-label"></label>
         <input asp-for="PIFirstName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="PIFirstName" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="PILastName" class="col-sm-3 col-form-label"></label>
        <input asp-for="PILastName" class="col-sm-9 form-control">
+       <span class="field-validation-valid text-danger" data-valmsg-for="PILastName" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="PIEmail" class="col-sm-3 col-form-label"></label>
         <input asp-for="PIEmail" class="col-sm-9 form-control">
-    </div>
+        <span class="field-validation-valid text-danger" data-valmsg-for="PIEmail" data-valmsg-replace="true"></span>
+    </div>    
     <div class="form-group row">
-        <label asp-for="PIEmailAdditional" class="col-sm-3 col-form-label"></label>
-        <input asp-for="PIEmailAdditional" class="col-sm-9 form-control">
-    </div><div class="form-group row">
         <label asp-for="PIFax" class="col-sm-3 col-form-label"></label>
         <input asp-for="PIFax" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="PIFax" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="PIPhone" class="col-sm-3 col-form-label"></label>
         <input asp-for="PIPhone" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="PIPhone" data-valmsg-replace="true"></span>
+    </div>
+    <div class="form-group row required">
+        <label asp-for="ResearcherFirstName" class="col-sm-3 col-form-label"></label>
+        <input asp-for="ResearcherFirstName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ResearcherFirstName" data-valmsg-replace="true"></span>
+    </div>
+    <div class="form-group row required">
+        <label asp-for="ResearcherLastName" class="col-sm-3 col-form-label"></label>
+        <input asp-for="ResearcherLastName" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ResearcherLastName" data-valmsg-replace="true"></span>
+    </div>
+    <div class="form-group row required">
+        <label asp-for="ResearcherEmail" class="col-sm-3 col-form-label"></label>
+        <input asp-for="ResearcherEmail" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ResearcherEmail" data-valmsg-replace="true"></span>
+    </div>
+    <div class="form-group row required">
+        <label asp-for="ResearcherPhone" class="col-sm-3 col-form-label"></label>
+        <input asp-for="ResearcherPhone" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ResearcherPhone" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="Address1" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address1" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address1" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="Address2" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address2" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address2" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="Address3" class="col-sm-3 col-form-label"></label>
         <input asp-for="Address3" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Address3" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="City" class="col-sm-3 col-form-label"></label>
         <input asp-for="City" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="City" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="State" class="col-sm-3 col-form-label"></label>
         <input asp-for="State" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="State" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row">
         <label asp-for="ZipCode" class="col-sm-3 col-form-label"></label>
         <input asp-for="ZipCode" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="ZipCode" data-valmsg-replace="true"></span>
     </div>
     <div class="form-group row required">
         <label asp-for="Country" class="col-sm-3 col-form-label"></label>
         <input asp-for="Country" class="col-sm-9 form-control">
+        <span class="field-validation-valid text-danger" data-valmsg-for="Country" data-valmsg-replace="true"></span>
     </div>
+    <div class="row">
+        <label asp-for="Notes" class="col-sm-3 col-form-label"></label>
+        <textarea asp-for="Notes" class="col-sm-9 form-control"></textarea>
+        <span class="field-validation-valid text-danger" data-valmsg-for="Notes" data-valmsg-replace="true"></span>
+    </div>
+    <div><hr></div>
     <div class="row">
         <input type="submit" value="Update" class="btn btn-primary"> | 
         <a asp-action="ShippingDetails" class="btn btn-info" asp-route-id="@Model.Id">Cancel</a> | 
         <a asp-action="Index" asp-area="client" asp-controller="Address" class="btn btn-info">Back to List</a>
-    </div>
+    </div>    
 </div>
 </form>
 
