@@ -5,12 +5,10 @@
 
 <h2>Create New Project</h2> 
 
-<form asp-action="CreateInspection">    
-    <div class="container col-lg-5">
-        <div class="row">
-             <div class="col-lg-12">
-                <div asp-validation-summary="All" class="text-danger"></div>
-            </div>
+<form asp-action="New">    
+    <div class="container col-lg-10">
+        <div class="row">             
+            <div asp-validation-summary="All" class="text-danger"></div>            
         </div>
         <div class="form-group row required">
             <label asp-for="Order.ProjectName" class="col-sm-3 col-form-label"></label>
@@ -21,7 +19,7 @@
             <label asp-for="Order.ShippingAddress" class="col-sm-3 col-form-label"></label>
             <select asp-for="Order.ShippingAddress" class="col-sm-3 form-control" asp-items="@(new SelectList(Model.ShippingAddresses, "Id", "AddressName"))"></select>                     
             <span class="field-validation-valid text-danger" data-valmsg-for="Order.ShippingAddress" data-valmsg-replace="true"></span> 
-            <span class="col-sm-6">We will deliver data to this address. If you do not see the shipping address you are looking for, please create it
+            <span class="col-sm-6">We will deliver data to this address. If you do not see the address you are looking for, please create it
                 <a asp-action="ShippingCreate" asp-area="client" asp-controller="Address">here</a></span>                    
         </div>
         <div class="form-group row required">
@@ -33,31 +31,31 @@
         </div>
         <div class="form-group row">
             <label asp-for="Order.PO" class="col-sm-3 col-form-label"></label>
-            <input type="checkbox" class="col-sm-3 form-control" asp-for="Order.PO">                                   
+            <input type="checkbox" class="col-sm-3 form-control" asp-for="Order.PO">  Is a PO being issued by our institution for this order?                                 
         </div>
         <div class="form-group row">
             <label asp-for="Order.PONumber" class="col-sm-3 col-form-label"></label>
-            <input type="text" class="col-sm-3 form-control" asp-for="Order.PONumber">                  
+            <input type="text" class="col-sm-3 form-control" asp-for="Order.PONumber"> 
+            <span class="field-validation-valid text-danger" data-valmsg-for="Order.PONumber" data-valmsg-replace="true"></span>                 
         </div>  
         <div class="form-group row required">
             <label asp-for="Order.PaymentMethod" class="col-sm-3 col-form-label"></label>
             <select asp-for="Order.PaymentMethod" class="col-sm-3 form-control" asp-items="@(new SelectList(Model.PaymentOptions))"></select>                     
-            <span class="field-validation-valid text-danger" data-valmsg-for="Order.ShippingAddress" data-valmsg-replace="true"></span> 
-            <span class="col-sm-6">We will deliver data to this address. If you do not see the shipping address you are looking for, please create it
-                <a asp-action="ShippingCreate" asp-area="client" asp-controller="Address">here</a></span>                    
+            <span class="field-validation-valid text-danger" data-valmsg-for="Order.ShippingAddress" data-valmsg-replace="true"></span>                             
         </div>
-        <div class="form-group row">
-            <label asp-for="Order.Hardcopy" class="col-sm-6 col-form-label">Invoices are delivered by e-mail from ekngo@ucdavis.edu. The SIF is going paperless.
-                Please check box if you require a hardcopy by post:
-            </label>
-            <input type="checkbox" asp-for="Order.Hardcopy" class="col-sm-3 form-control">
+        <div class="row">
+            Invoices are delivered by e-mail from sifaccounting@ucdavis.edu. 
         </div>
         <div class="form-group row">
             <label asp-for="Order.OrderComments" class="col-sm-3 col-form-label"></label>
             <textarea class="form-control" asp-for="Order.OrderComments" rows="3"></textarea>
         </div> 
         <div class="row">
-            <input type="submit" value="Create Inspection" class="btn btn-primary"></input> | 
+            By submitting this request, you acknowledge the following Terms & Conditions: The university, its officers, employees, and agents shall not be accountable for any loss, expense (including attorneys’ fees), damage, or liability of any kind resulting from or arising out of services provided hereunder unless caused by negligent or willful acts or omissions by the university, its officers, employees, or agents.
+
+        </div>
+        <div class="row">
+            <input type="submit" value="Begin Order" class="btn btn-primary"></input> | 
             <a class="btn btn-dark" asp-action="Index" asp-controller="Order" asp-area="Client">Cancel</a>
         </div>
     </div>
