@@ -1,6 +1,6 @@
 @model SIFCore.Models.AccountViewModels.LoginViewModel
 
-
+<h3>Note: password reset options will be available.</h3>
 <h2>Login</h2>
 <div class="row">
     <div class="col-md-4">
